Refactoring + addressing comments
