+++ components/org.wso2.carbon.event.outbound.adapter.websubhub/src/main/java/org/wso2/carbon/event/outbound/adapter/websubhub/util/WebSubHubEventAdapterConstants.java
@@ -16,17 +16,15 @@
  * under the License.
  */
 
-package org.wso2.carbon.event.outbound.adapter.websubhub.internal.util;
+package org.wso2.carbon.event.outbound.adapter.websubhub.util;
 
 /**
  * Keep constants required by the WebSubHub Event Adapter.
  */
 public class WebSubHubEventAdapterConstants {
-
     private WebSubHubEventAdapterConstants() {
 
     }
-
     public static final String ADAPTER_MESSAGE_TENANT_DOMAIN = "tenant-domain";
     public static final String EVENT_ISSUER = "Asgardeo";
     public static final String AUDIENCE_BASE_URL = "https://websubhub/topics/";
