+++ pom.xml
@@ -16,7 +16,8 @@
   ~ specific language governing permissions and limitations
   ~ under the License.
   -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
     <parent>
         <groupId>org.wso2</groupId>
         <artifactId>wso2</artifactId>
@@ -30,7 +31,7 @@
 
     <packaging>pom</packaging>
     <name>Outbound Adapters</name>
-    <description>This is to maintain a set of output event adapters which do not depend on Siddhi event stream techniques.</description>
+    <description>Maintain output event adapters which do not depend on Siddhi event stream techniques.</description>
     <url>http://wso2.org</url>
 
     <modules>
@@ -39,7 +40,8 @@
 
     <scm>
         <url>https://github.com/wso2-extensions/identity-outbound-adapters.git</url>
-        <developerConnection>scm:git:https://github.com/wso2-extensions/identity-outbound-adapters.git</developerConnection>
+        <developerConnection>scm:git:https://github.com/wso2-extensions/identity-outbound-adapters.git
+        </developerConnection>
         <connection>scm:git:https://github.com/wso2-extensions/identity-outbound-adapters.git</connection>
         <tag>HEAD</tag>
     </scm>
@@ -93,33 +95,6 @@
         <pluginManagement>
             <plugins>
                 <plugin>
-                    <groupId>org.codehaus.mojo</groupId>
-                    <artifactId>buildnumber-maven-plugin</artifactId>
-                    <version>${maven.buildnumber.plugin.version}</version>
-                    <executions>
-                        <execution>
-                            <phase>validate</phase>
-                            <goals>
-                                <goal>create</goal>
-                            </goals>
-                        </execution>
-                    </executions>
-                    <configuration>
-                        <doCheck>false</doCheck>
-                        <doUpdate>false</doUpdate>
-                    </configuration>
-                </plugin>
-                <plugin>
-                    <groupId>org.apache.maven.plugins</groupId>
-                    <artifactId>maven-surefire-plugin</artifactId>
-                    <version>${maven.surefire.plugin.version}</version>
-                </plugin>
-                <plugin>
-                    <groupId>org.jacoco</groupId>
-                    <artifactId>jacoco-maven-plugin</artifactId>
-                    <version>${jacoco.version}</version>
-                </plugin>
-                <plugin>
                     <groupId>org.apache.felix</groupId>
                     <artifactId>maven-bundle-plugin</artifactId>
                     <version>${maven.bundle.plugin.version}</version>
@@ -158,6 +133,26 @@
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>buildnumber-maven-plugin</artifactId>
+                <version>${maven.buildnumber.plugin.version}</version>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-javadoc-plugin</artifactId>
+                <configuration>
+                    <source>1.8</source>
+                </configuration>
+                <executions>
+                    <execution>
+                        <id>attach-javadocs</id>
+                        <goals>
+                            <goal>jar</goal>
+                        </goals>
+                        <configuration>
+                            <!--This parameter disables doclint-->
+                            <doclint>none</doclint>
+                        </configuration>
+                    </execution>
+                </executions>
             </plugin>
         </plugins>
     </build>
@@ -174,5 +169,15 @@
         <jackson.version>2.4.4</jackson.version>
         <com.googlecode.json-simple.wso2.version>1.1.wso2v1</com.googlecode.json-simple.wso2.version>
         <identity.outbound.adapters.version>1.0.0-SNAPSHOT</identity.outbound.adapters.version>
+        <findbug-maven-plugin.version>3.0.5</findbug-maven-plugin.version>
+
+        <commons-lang.wso2.osgi.version.range>[2.6.0,3.0.0)</commons-lang.wso2.osgi.version.range>
+        <org.apache.commons.logging.range>[1.2.0,2.0.0)</org.apache.commons.logging.range>
+        <httpasyncclient.version.range>[4.0.0, 5.0.0)</httpasyncclient.version.range>
+        <osgi.framework.imp.pkg.version.range>[1.7.0, 2.0.0)</osgi.framework.imp.pkg.version.range>
+        <osgi.service.component.imp.pkg.version.range>[1.2.0, 2.0.0)</osgi.service.component.imp.pkg.version.range>
+        <carbon.identity.framework.imp.pkg.version.range>[5.20.90, 6.0.0)
+        </carbon.identity.framework.imp.pkg.version.range>
+        <com.googlecode.json-simple.wso2.version.range>[0.0.0, 2.0.0)</com.googlecode.json-simple.wso2.version.range>
     </properties>
 </project>
